add ObservableCounterSlice logic into a different file
